SCRUM-41 implement financial REST API

+++ src/test/setup.ts
@@ -1,5 +1,7 @@
 import '@testing-library/jest-dom/vitest'
 
 beforeEach(() => {
-  localStorage.clear()
+  if (typeof localStorage !== 'undefined') {
+    localStorage.clear()
+  }
 })
